requirements.txt
streamlit
openai
audio-recorder-streamlit
psycopg2-binary
sqlalchemy